clarifiy: Requirements
- putting the lower layer requirements in the descriptive text before
stating the actual requirements.
The text said these were out of scope, yet still, it was a requirements.
That was confusing.

- adding comments showing which parts of DNS-SD (or of a
  hypothetical badly designed privacy extension) violate which
requirement.

- DNS-SD clients to not publish their hostname or IP addresses via
DNS-SD. Using mDNS, each device acts as both client and server.
The server part discloses this information.
Maybe we should clarify that point as well.

- private servers should also not disclose service types

- Not disclosing host names is not a separate requirement.
It is now a part of (1).

- Change the "IP address requirement" to only cover IP addresses
disclosed via DNS-SD messages